odoo_tutorial/movie_theater_realtime/pyproject.toml
[project]
name = "movie-theater-realtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymongo",
    "python-dotenv",
    "pydantic",
    "fastapi",
    "uvicorn",
]
